tracks deleted panes. prevents dom from rerendering when a pane is deleted. use if statements to remove deleted panes from the dom!!!

+++ src/routes/+page.svelte
@@ -14,6 +14,7 @@
 	});
 
 	let elements = $state([]);
+	let deletedElements = $state({});
 
 	function onGridUpdate() {
 		let gta = renderGridTemplateAreas(root);
@@ -30,18 +31,20 @@
 			grid += '"' + s + '"\n';
 		}
 
-		elements = Object.keys(els).sort((a: string, b: string) => {
-			let aval = 0,
-				bval = 0;
+		elements = Object.keys(els)
+			.concat(Object.keys(deletedElements))
+			.sort((a: string, b: string) => {
+				let aval = 0,
+					bval = 0;
 
-			for (let i = 0; i < a.length; i++) {
-				aval += a.charCodeAt(i) - 96;
-			}
-			for (let i = 0; i < b.length; i++) {
-				bval += b.charCodeAt(i) - 96;
-			}
-			return aval - bval;
-		});
+				for (let i = 0; i < a.length; i++) {
+					aval += a.charCodeAt(i) - 96;
+				}
+				for (let i = 0; i < b.length; i++) {
+					bval += b.charCodeAt(i) - 96;
+				}
+				return aval - bval;
+			});
 
 		template = `display: grid;
         grid-template-columns: repeat(${gta.length}, ${gta[0].length});
@@ -141,6 +144,7 @@
 		}
 
 		if (found) {
+			deletedElements[n.left.id] = n.left.id;
 			//do delete. this is the parent
 			if (n.right.split) {
 				n.split = n.right.split;
@@ -152,6 +156,7 @@
 				n.left = undefined;
 				n.right = undefined;
 			}
+
 			onGridUpdate();
 			return;
 		}
@@ -161,6 +166,7 @@
 		}
 
 		if (found) {
+			deletedElements[n.right.id] = n.right.id;
 			//do delete this is the parent
 			if (n.left.split) {
 				n.split = n.left.split;
@@ -172,6 +178,7 @@
 				n.left = undefined;
 				n.right = undefined;
 			}
+
 			onGridUpdate();
 			return;
 		}
@@ -185,41 +192,48 @@
 
 <div class="flex h-[100vh] w-full flex-col">
 	{JSON.stringify(root)}
+	{JSON.stringify(deletedElements)}
 	<div style="min-width: 1px; {template}" class="h-[100%] w-full">
 		<!-- svelte-ignore a11y_click_events_have_key_events -->
 		{#each elements as a, idx}
 			<!-- svelte-ignore a11y_click_events_have_key_events -->
 			<!-- svelte-ignore a11y_no_static_element_interactions -->
-			<div
-				style="grid-area: {a};"
-				class="header w-full items-center border border-neutral-200 bg-neutral-950 text-center"
-			>
-				<div class="flex h-full flex-col items-center justify-center text-neutral-200">
-					{a}
-					{new Date().getSeconds()}
-					<br />
-					<button
-						class="bg-green-500 text-neutral-700"
-						onclick={() => {
-							splitPane(a, 'v');
-						}}>Vertical Split</button
-					>
-					<button
-						class="bg-green-500 text-neutral-700"
-						onclick={() => {
-							splitPane(a, 'h');
-						}}>Horizontal</button
-					>
+			{#if !deletedElements[a]}
+				<div
+					style="grid-area: {a};"
+					class="header w-full items-center border border-neutral-200 bg-neutral-950 text-center {deletedElements[
+						a
+					]
+						? 'hidden'
+						: ''}"
+				>
+					<div class="flex h-full flex-col items-center justify-center text-neutral-200">
+						{a}
+						{new Date().getSeconds()}
+						<br />
+						<button
+							class="bg-green-500 text-neutral-700"
+							onclick={() => {
+								splitPane(a, 'v');
+							}}>Vertical Split</button
+						>
+						<button
+							class="bg-green-500 text-neutral-700"
+							onclick={() => {
+								splitPane(a, 'h');
+							}}>Horizontal</button
+						>
 
-					<button
-						class="bg-green-500 text-neutral-700"
-						onclick={() => {
-							console.log('root', root);
-							deletePane(root, a);
-						}}>Delete</button
-					>
+						<button
+							class="bg-green-500 text-neutral-700"
+							onclick={() => {
+								console.log('root', root);
+								deletePane(root, a);
+							}}>Delete</button
+						>
+					</div>
 				</div>
-			</div>
+			{/if}
 		{/each}
 	</div>
 </div>
